webjobs/user_data_sync/requirements.txt
adal=0.4.7
cryptography==2.1.4
pyopenssl==17.5.0
